Move all serialization functions to `serialize.go`.
